[full ci] Rename vicbackends package to backends
Go convention is to name the directory the same as the pacakge.  Also
vicbackends stutters.
