+++ lib/apiservers/engine/backends/image.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package vicbackends
+package backends
 
 import (
 	"fmt"
